Move FolderCompletion and FolderSummary declarations to events.
Rename to avoid conflicts with the event type constants of the same
name.  Use local type aliases in the producing packages, avoiding
explicit references to the events package everywhere.
